homework 6 - review comments fixed
