+++ src/test/java/hw6/RunTests.java
@@ -1,8 +1,8 @@
-package hw6.ex2;
+package hw6;
 
 import cucumber.api.CucumberOptions;
 import cucumber.api.testng.AbstractTestNGCucumberTests;
 
-@CucumberOptions(features = "classpath:hw6/ex2")
+@CucumberOptions(features = "classpath:hw6")
 public class RunTests extends AbstractTestNGCucumberTests {
 }
