--- src/test/java/hw6/ex1/RunTests.java
@@ -1,8 +1,0 @@
-package hw6.ex1;
-
-import cucumber.api.CucumberOptions;
-import cucumber.api.testng.AbstractTestNGCucumberTests;
-
-@CucumberOptions(features = "classpath:hw6/ex1")
-public class RunTests extends AbstractTestNGCucumberTests {
-}
