--- src/test/java/hw6/ex1/steps/FirstExerciseModalSteps.java
@@ -1,15 +1,0 @@
-package hw6.ex1.steps;
-
-import cucumber.api.java.en.Given;
-import hw6.HomePage;
-import hw6.common.AbstractBaseSteps;
-import hw6.common.WebDriverSingleton;
-
-public class FirstExerciseModalSteps extends AbstractBaseSteps {
-
-    @Given("^I open EPAM JDI Site$")
-    public void iOpenEpamJdiSite() {
-        homePage = new HomePage(WebDriverSingleton.INSTANCE.getDriver());
-        homePage.open();
-    }
-}
