+++ src/test/java/hw6/hooks/CucumberHooks.java
@@ -1,4 +1,4 @@
-package hw6.ex2;
+package hw6.hooks;
 
 import cucumber.api.java.After;
 import cucumber.api.java.Before;
@@ -9,7 +9,7 @@ import org.openqa.selenium.chrome.ChromeDriver;
 
 import java.util.concurrent.TimeUnit;
 
-public class Ex2CucumberHooks {
+public class CucumberHooks {
 
     @Before
     public void setUp() {
