+++ src/test/java/hw6/steps/AssertionSteps.java
@@ -1,17 +1,21 @@
-package hw6.ex1.steps;
+package hw6.steps;
 
+import cucumber.api.DataTable;
 import cucumber.api.java.en.Then;
 import hw6.common.AbstractBaseSteps;
+import hw6.common.TestStorage;
 import hw6.enums.IconItemsText;
+import org.testng.asserts.SoftAssert;
 
 import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertTrue;
 
-public class FirstExerciseAssertionSteps extends AbstractBaseSteps {
+public class AssertionSteps extends AbstractBaseSteps {
 
     @Then("^username should be '([^\"]+)' on the Home page$")
     public void usernameShouldBeOnTheHomePage(String expectedUsername) {
@@ -98,5 +102,67 @@ public class FirstExerciseAssertionSteps extends AbstractBaseSteps {
     @Then("^log row should be '([^\"]+)' on the Different Elements page$")
     public void logRowShouldBeDisplayedOnTheDifferentElementsPage(final String logRecord) {
         assertTrue(differentElementsPage.getLastLogRecordText().contains(logRecord));
+    }
+
+    @Then("^'([^\"]+)' page is opened$")
+    public void pageIsOpened(final String pageTitle) {
+        assertEquals(userTablePage.getPageTitle(), pageTitle);
+    }
+
+    @Then("^6 NumberType Dropdowns should be displayed on Users Table on User Table Page$")
+    public void numbersTypeDropdownsShouldBeDisplayedOnUsersTableOnUserTablePage() {
+        userTablePage.getNumberTypeDropdowns().forEach(t -> assertTrue(t.isDisplayed()));
+    }
+
+    @Then("^6 User names should be displayed on Users Table on User Table Page$")
+    public void useNamesShouldBeDisplayedOnUsersTableOnUserTablePage() {
+        userTablePage.getUserNamesList().forEach(t -> t.isDisplayed());
+    }
+
+    @Then("^6 Description images should be displayed on Users Table on User Table Page$")
+    public void descriptionImagesShouldBeDisplayedOnUsersTableOnUserTablePage() {
+        userTablePage.getDescriptionImages().forEach(t -> t.isDisplayed());
+
+    }
+
+    @Then("^6 Description texts under images should be displayed on Users Table on User Table Page$")
+    public void descriptionTextsUnderImagesShouldBeDisplayedOnUsersTableOnUserTablePage() {
+        userTablePage.getDescriptionText().forEach(t -> t.isDisplayed());
+    }
+
+    @Then("^6 checkboxes should be displayed on Users Table on User Table Page$")
+    public void checkboxesShouldBeDisplayedOnUsersTableOnUserTablePage() {
+        userTablePage.getCheckboxes().forEach(t -> t.isDisplayed());
+
+    }
+
+    @Then("^User table should contain following values:$")
+    public void userTableShouldContainFollowingValues(DataTable table) {
+        SoftAssert sa = new SoftAssert();
+        List<Map<String, String>> values = table.asMaps(String.class, String.class);
+
+        values.forEach(t -> sa.assertTrue(
+                userTablePage.getUserNumberListText().contains(t.get("Number")),
+                "User Table numbers checker"
+        ));
+        values.forEach(t -> sa.assertTrue(
+                userTablePage.getUserNamesText().contains(t.get("User")),
+                "User Table userNames checker"
+        ));
+        values.forEach(t -> sa.assertTrue(
+                userTablePage.getUserDescriptionText().contains(t.get("Description")),
+                "User Table userDescriptions checker"
+        ));
+        sa.assertAll();
+    }
+
+    @Then("^log row '([^\"]+)' text should be displayed in log section$")
+    public void logRowTextShouldBeDisplayedInLogSection(final String logRecord) {
+        assertTrue(userTablePage.getUserTableLogRecordText().contains(logRecord));
+    }
+
+    @Then("^dropdown list should contain values$")
+    public void dropdownListShouldContainValues(List<String> expectedOptionsList) {
+        assertTrue(expectedOptionsList.containsAll(TestStorage.INSTANCE.getOptionsList()));
     }
 }
