+++ src/test/java/hw6/steps/ModalSteps.java
@@ -1,14 +1,14 @@
-package hw6.ex2.steps;
+package hw6.steps;
 
 import cucumber.api.java.en.Given;
 import hw6.HomePage;
 import hw6.common.AbstractBaseSteps;
 import hw6.common.WebDriverSingleton;
 
-public class SecondExerciseModalSteps extends AbstractBaseSteps {
+public class ModalSteps extends AbstractBaseSteps {
 
-    @Given("^I am on Home Page$")
-    public void iAmOnHomePage() {
+    @Given("^I open EPAM JDI Site$")
+    public void iOpenEpamJdiSite() {
         homePage = new HomePage(WebDriverSingleton.INSTANCE.getDriver());
         homePage.open();
     }
